+++ src/frameworks/native/device.rs
@@ -45,12 +45,12 @@ impl IDevice for Cpu {
 
     fn sync_in(&self, source: &DeviceType, source_data: &MemoryType, dest_data: &mut FlatBox) -> Result<(), DeviceError> {
         match source {
-            &DeviceType::Native(ref cpu) => unimplemented!(),
-            &DeviceType::OpenCL(ref context) => unimplemented!(),
+            &DeviceType::Native(_) => unimplemented!(),
+            &DeviceType::OpenCL(_) => unimplemented!(),
             #[cfg(feature = "cuda")]
             &DeviceType::Cuda(ref context) => {
                 match source_data.as_cuda() {
-                    Some(mut h_mem) => Ok(try!(context.sync_out(&DeviceType::Native(self.clone()), &h_mem, dest_data))),
+                    Some(h_mem) => Ok(try!(context.sync_out(&DeviceType::Native(self.clone()), &h_mem, dest_data))),
                     None => Err(DeviceError::Native(Error::Memory("Expected CUDA Memory")))
                 }
             },
